DarkKnight Locomotion 수정 완료(BlendSpace1D 사용)

+++ UnrealRPG/Source/UnrealRPG/DarkKnight.h
@@ -69,7 +69,7 @@ protected:
 	void ChangeCombatState(ECombatState NextCombatState);
 
 	UFUNCTION(BlueprintCallable)
-	void FaceOff(int32 NextWalkDirection);
+	void FaceOff(float NextWalkDirection);
 
 	UFUNCTION(BlueprintCallable)
 	void EndFaceOff();
